Application/requirements.txt
Flask==2.0.1
Werkzeug==2.0.2
python-dotenv==0.19.2
requests==2.25.1
flask-socketio==5.3.6